pyproject.toml
[project]
name = "orchestra"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Amitai Sopher", email = "amitaisopher@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "aws-cdk-lib==2.215.0",
    "boto3>=1.40.55",
    "constructs>=10.0.0,<11.0.0",
    "dotenv>=0.9.9",
]

[project.scripts]
orchestra = "orchestra:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "ruff>=0.14.1",
]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "UP", "B", "ANN", "COM"]
